Move legacy testsupport into but-testsupport
