+++ crates/but-testsupport/src/fixtures/user/minimal.v1
