+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -64,8 +64,7 @@ but-schemars = { workspace = true, optional = true }
 [dev-dependencies]
 pretty_assertions = "1.4"
 but-graph.workspace = true
-but-testsupport.workspace = true
-gitbutler-testsupport.workspace = true
+but-testsupport = { workspace = true, features = ["legacy"] }
 gitbutler-workspace.workspace = true
 gix = { workspace = true, features = [] }
 glob = "0.3.3"
