+++ crates/gitbutler-branch-actions/tests/branch-actions/virtual_branches/amend.rs
@@ -1,7 +1,7 @@
 use but_core::{DiffSpec, HunkHeader};
 use but_oxidize::{ObjectIdExt, OidExt};
+use but_testsupport::legacy::stack_details;
 use gitbutler_branch::BranchCreateRequest;
-use gitbutler_testsupport::stack_details;
 
 use super::{list_commit_files, *};
