+++ crates/gitbutler-branch-actions/tests/branch-actions/virtual_branches/init.rs
@@ -1,4 +1,4 @@
-use gitbutler_testsupport::stack_details;
+use but_testsupport::legacy::stack_details;
 
 use super::*;
