+++ crates/gitbutler-branch-actions/tests/branch-actions/virtual_branches/save_and_unapply_virtual_branch.rs
@@ -1,6 +1,6 @@
 use but_core::DiffSpec;
 use but_hunk_assignment::HunkAssignmentRequest;
-use gitbutler_testsupport::stack_details;
+use but_testsupport::legacy::stack_details;
 
 use super::*;
