+++ crates/gitbutler-branch-actions/tests/branch-actions/virtual_branches/unapply_without_saving_virtual_branch.rs
@@ -1,4 +1,4 @@
-use gitbutler_testsupport::stack_details;
+use but_testsupport::legacy::stack_details;
 
 use super::*;
