+++ crates/gitbutler-project/src/project.rs
@@ -192,10 +192,10 @@ impl Project {
 }
 
 /// Testing
-// TODO: remove once `gitbutler-testsupport` isn't needed anymore, and `gitbutler-repo`
+// TODO: remove once the remaining legacy testsupport constructor isn't needed anymore, and `gitbutler-repo`
 impl Project {
     /// A special constructor needed as `worktree_dir` isn't accessible anymore.
-    pub fn new_for_gitbutler_testsupport(title: String, worktree_dir: PathBuf) -> Self {
+    pub fn new_for_but_testsupport(title: String, worktree_dir: PathBuf) -> Self {
         let project_id = ProjectHandleOrLegacyProjectId::ProjectHandle(
             ProjectHandle::from_path(&worktree_dir)
                 .expect("testsupport projects require a valid path for ProjectHandle"),
