--- crates/gitbutler-testsupport/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "gitbutler-testsupport"
-version = "0.0.0"
-edition.workspace = true
-publish = false
-rust-version.workspace = true
-
-[lib]
-path = "src/lib.rs"
-doctest = false
-test = false
-
-[dependencies]
-but-core = { workspace = true, features = ["legacy"] }
-but-settings.workspace = true
-but-oxidize.workspace = true
-but-workspace.workspace = true
-but-fs = {workspace = true, features = ["legacy"]}
-but-ctx.workspace = true
-but-project-handle.workspace = true
-
-gitbutler-branch-actions = { path = "../gitbutler-branch-actions" }
-gitbutler-repo = { path = "../gitbutler-repo" }
-gitbutler-project.workspace = true
-gitbutler-user.workspace = true
-gitbutler-reference.workspace = true
-gitbutler-url.workspace = true
-gitbutler-stack.workspace = true
-
-anyhow.workspace = true
-git2.workspace = true
-tempfile.workspace = true
-keyring.workspace = true
-serde_json.workspace = true
-gix-testtools.workspace = true
-gix.workspace = true
-termtree = "0.5.1"
-uuid.workspace = true
-
-[lints]
-workspace = true
