+++ etc/plans/git2-to-gix-migration.md
@@ -280,9 +280,10 @@ Target tests/modules:
 - `crates/gitbutler-repo/tests/create_wd_tree.rs`
 - `crates/gitbutler-repo/tests/managed_hooks_tests.rs`
 - `crates/gitbutler-project/tests/project/main.rs`
-- `crates/gitbutler-testsupport/src/lib.rs`
-- `crates/gitbutler-testsupport/src/suite.rs`
-- `crates/gitbutler-testsupport/src/test_project.rs`
+- `crates/but-testsupport/src/legacy.rs`
+- `crates/but-testsupport/src/legacy/suite.rs`
+- `crates/but-testsupport/src/legacy/test_project.rs`
+- `crates/but-testsupport/src/legacy/testing_repository.rs`
 
 Tasks:
 
@@ -306,8 +307,9 @@ Current reconciliation notes:
 - `cargo metadata` still reports 18 crates with a normal `git2` dependency, so manifest cleanup remains blocked on production and test/helper cleanup.
 - `crates/gitbutler-stack/tests/mod.rs` now bakes stack commit history via `gix` traversal instead of `ctx.git2_repo`.
 - `crates/gitbutler-branch-actions/tests/branch-actions/virtual_branches/workspace_migration.rs` now inspects HEAD through `gix`.
-- `crates/gitbutler-edit-mode/tests/edit_mode.rs`, `crates/gitbutler-repo/tests/create_wd_tree.rs`, `crates/gitbutler-repo/tests/managed_hooks_tests.rs`, `crates/gitbutler-project/tests/project/main.rs`, and `crates/gitbutler-testsupport/src/*` still contain substantial direct `git2` usage.
-- `gitbutler-testsupport` should be treated as a migration target, not a keeper crate: move coverage helpers to `but-testsupport`, then delete `gitbutler-testsupport`.
+- On 2026-03-21, the shared legacy helper surface (`Case`, `Suite`, `TestProject`, `testing_repository`, `secrets`, `paths`, `stack_details`, and related setup helpers) was moved into `crates/but-testsupport/src/legacy/*`, and the current test users in `gitbutler-operating-modes`, `gitbutler-project`, `gitbutler-repo`, and `gitbutler-branch-actions` were switched from `gitbutler-testsupport` to `but-testsupport` with the `legacy` feature.
+- `crates/gitbutler-edit-mode/tests/edit_mode.rs`, `crates/gitbutler-repo/tests/create_wd_tree.rs`, `crates/gitbutler-repo/tests/managed_hooks_tests.rs`, `crates/gitbutler-project/tests/project/main.rs`, and `crates/but-testsupport/src/legacy/*` still contain substantial direct `git2` usage.
+- On 2026-03-21, the `gitbutler-testsupport` shim crate was deleted after the remaining users had moved to `but-testsupport`.
 - `crates/gitbutler-branch-actions/tests/branch-actions/squash.rs` is currently the largest remaining `ctx.git2_repo` consumer in tests.
 - Remaining direct `ctx.git2_repo` usage in tests should trend toward hard-boundary coverage and fixture/setup helpers only as reopened production migrations land.
 
@@ -359,8 +361,8 @@ Current execution status:
 - Workstreams C, D, E, and G remain active under the narrowed hard boundary.
 - Workstream F has not started yet.
 - Residual `git2` usage is not yet limited to the hard boundary and compatibility/public-boundary crates.
-- Open actionable items: repo adapter cleanup, transport/auth cleanup, workspace/oplog boundary extraction, and follow-on test/helper migration including `gitbutler-testsupport` replacement.
-- Recommended next wave: Wave 5, starting with `crates/gitbutler-workspace/src/branch_trees.rs`, `crates/gitbutler-oplog/src/oplog.rs`, and `gitbutler-testsupport` helper migration into `but-testsupport`.
+- Open actionable items: repo adapter cleanup, transport/auth cleanup, workspace/oplog boundary extraction, and continued test/helper `git2` reduction inside `but-testsupport::legacy`.
+- Recommended next wave: Wave 5, starting with `crates/gitbutler-workspace/src/branch_trees.rs`, `crates/gitbutler-oplog/src/oplog.rs`, and cleanup of the remaining `git2`-heavy helpers/tests that were just consolidated into `but-testsupport`.
 
 ## Acceptance Criteria
 
@@ -399,8 +401,8 @@ Remaining direct `git2` usage after reconciliation is split between hard-boundar
 - Hard-boundary and boundary-adjacent runtime code: `crates/but-core/src/worktree/checkout/*`, `crates/gitbutler-workspace/src/branch_trees.rs`, `crates/gitbutler-edit-mode/src/lib.rs` (treated as a boundary crate for its remaining checkout/index handoff), `crates/gitbutler-oplog/src/oplog.rs`, and the `git2` index/reset portions of `crates/gitbutler-branch-actions/src/integration.rs`
 - Actionable repo/transport/frontend adapters: `crates/gitbutler-repo/src/repository_ext.rs`, `crates/gitbutler-repo/src/commands.rs`, `crates/gitbutler-repo/src/rebase.rs`, `crates/gitbutler-repo/src/credentials.rs`, `crates/gitbutler-repo/src/hooks.rs`, `crates/gitbutler-repo/src/managed_hooks.rs`, `crates/gitbutler-repo/src/remote.rs`, `crates/gitbutler-repo/src/staging.rs`, `crates/gitbutler-repo-actions/src/repository.rs`, `crates/gitbutler-tauri/src/projects.rs`, and `crates/but-napi/src/lib.rs`
 - Foundational/shared compatibility boundaries still to shrink once callers move: `crates/but-ctx/src/lib.rs`, `crates/but-oxidize/src/lib.rs`, `crates/but-serde/src/lib.rs`, `crates/but-schemars/src/lib.rs`, `crates/gitbutler-project/src/lib.rs`, `crates/gitbutler-repo/src/lib.rs`, `crates/gitbutler-cherry-pick/src/*`, `crates/gitbutler-commit/src/commit_ext.rs`, and the remaining `git2` type boundary in `crates/gitbutler-stack/src/stack.rs`
-- Legacy test and fixture/setup code that still constructs `git2` values directly: selected `crates/gitbutler-branch-actions/tests/*`, `crates/gitbutler-edit-mode/tests/edit_mode.rs`, `crates/gitbutler-repo/tests/*`, `crates/gitbutler-project/tests/project/main.rs`, and `crates/gitbutler-testsupport/src/*`
-- Planned helper consolidation: migrate remaining reusable fixture/test utilities from `gitbutler-testsupport` into `but-testsupport`, then remove `gitbutler-testsupport` from the tree.
+- Legacy test and fixture/setup code that still constructs `git2` values directly: selected `crates/gitbutler-branch-actions/tests/*`, `crates/gitbutler-edit-mode/tests/edit_mode.rs`, `crates/gitbutler-repo/tests/*`, `crates/gitbutler-project/tests/project/main.rs`, and `crates/but-testsupport/src/legacy/*`
+- Planned helper retirement: keep shrinking `git2` inside `but-testsupport::legacy` until the legacy feature itself can collapse away.
 
 Residual `git2` usage that is currently consistent with the narrowed hard boundary:
